Move recipe to framework and include it in framework 6 migration
